Added an internal field to store the original JSON string
This is used instead of the conversion from objecto->json when using the method ToJsonString.

There is an important issue with this class: the uuid field on the cancellation response is not being correctly parsed, because the field is a dynamic value (the UUID) so the RTTI can't map it to the object.
Adding the original JSON string allows the user of this class to parse the JSON directly and extract those values if needed.