Removed the storing of the CFDI into a "temporary" file
Storing the CFDI XML into a file on the current directory is not adequate in a lot of cases: if the user doesn't have privileges to modify that folder, or there are several concurrent users that might need to use this service, and so on.
It's not needed to do that file, in the first place: you can pass the XML directly using a TStringStream.