Feature/ss879 (#22) (#23)
* Removed unnecessary units from the uses clauses, and moved what was possible to the uses inside the implementation section

This to avoid issues when generating C++Builder compatible code, as the extra includes generate compiling problems with big projects

* Cleanup of unused variable

* Added an internal field to store the original JSON string

This is used instead of the conversion from objecto->json when using the method ToJsonString.

There is an important issue with this class: the uuid field on the cancellation response is not being correctly parsed, because the field is a dynamic value (the UUID) so the RTTI can't map it to the object.
Adding the original JSON string allows the user of this class to parse the JSON directly and extract those values if needed.

* Removed the storing of the CFDI into a "temporary" file

Storing the CFDI XML into a file on the current directory is not adequate in a lot of cases: if the user doesn't have privileges to modify that folder, or there are several concurrent users that might need to use this service, and so on.
It's not needed to do that file, in the first place: you can pass the XML directly using a TStringStream.

* Improved response handling to retrieve UUID statuses

* Improve form to receive response attributes from cancellation methods

* Improve form to receive response attributes from cancellation methods

* Added attribute invocation specifications and precision to support libraries for TLS 1.2 compatibility

* add LICENSE file for Luna Soft project (GPL v3)

* add dll files to configuration

* remove comments

* update version

---------

Co-authored-by: martinfnsw <88680430+martinfnsw@users.noreply.github.com>
Co-authored-by: Rodrigo Gómez <rgomez@trafficsystem.com.mx>